Fixed #70: change requested formats
Requested format for videos was `webm` or `mp4` based on user selection.
This used to be fine as most videos would return a usable video for both format or
fallback and convert to the requested format.
Recently, Youtube changed and seems to not offer video with audio option for `webm`.
As `webm` is also used for audio, Youtube-dl would consider this a video-only webm
request, given the available formats.
This resulted in all videos being silently without audio.

This fixes it by specifically requesting the best available audio+video for `format` and if
not available would request and merge separate files for both and if still not, would
request the best available video+audio and convert it as before